foundry simple storage: docs

+++ src/SimpleStorage.sol
@@ -3,13 +3,13 @@
 pragma solidity 0.8.19;
 
 contract SimpleStorage {
-    uint num;
+    uint256 num;
 
-    function store(uint _num) public {
+    function store(uint256 _num) public {
         num = _num;
     }
 
-    function get() public view returns (uint) {
+    function get() public view returns (uint256) {
         return num;
     }
 }
